pyproject.toml
[project]
name = "muse-python"
version = "0.1.0"
description = "Test project to experiment using Python with Muse"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cowsay>=6.1",
]

[project.scripts]
deploy = "muse-python:deploy:main"

[dependency-groups]
dev = [
    "python-dotenv>=1.0.1",
]
